requirements.txt
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fonttools==4.57.0
frozendict==2.4.6
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
load-dotenv==0.1.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
yfinance==0.2.55
